Type: Feat Body: Feat
Footer:
calculateExitCost 함수 임시 완성
연 수익률로 계산 편의성 고려
청구서로 인한 감액 제외

Related to: #25

+++ src/main/java/com/shinhan/peoch/lifecycleincome/service/ExitCostService.java
@@ -24,7 +24,7 @@ public class ExitCostService {
     @Autowired
     PaymentRepository paymentRepository;
 
-    public ExitResponseDTO exitResponseService(Integer userId){
+    public ExitResponseDTO exitResponseService(long userId){
         InvestmentEntity firstExpectedIncomeEntity = investmentRepository.findFirstByUserIdOrderByUpdatedAtDesc(userId);
         InvestmentEntity lastExpectedIncomeEntity = investmentRepository.findFirstByUserIdOrderByCreatedAtAsc(userId);
         ExitResponseDTO exitResponseDTO= ExitResponseDTO.builder()
