+++ src/main/java/com/shinhan/repository/InvestmentRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 public interface InvestmentRepository extends JpaRepository<InvestmentEntity, Integer> {
 
     // 최신 updatedAt 기준으로 가장 최근 1개 엔티티 조회
-    InvestmentEntity findFirstByUserIdOrderByUpdatedAtDesc(Integer userId);
+    InvestmentEntity findFirstByUserIdOrderByUpdatedAtDesc(long userId);
 
     // 가장 오래된 createdAt 기준으로 가장 오래된 1개 엔티티 조회
-    InvestmentEntity findFirstByUserIdOrderByCreatedAtAsc(Integer userId);
+    InvestmentEntity findFirstByUserIdOrderByCreatedAtAsc(long userId);
 }
